refactored tests successive
